server isn't affected by other signals

+++ server.c
@@ -6,12 +6,13 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/29 19:23:02 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/06/29 19:35:25 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/07/01 16:00:20 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/include/ft_printf.h"
 #include "minitalk.h"
+#include <sys/signal.h>
 #include <unistd.h>
 
 void	handler(int sig, siginfo_t *client, void *con)
@@ -29,7 +30,8 @@ void	handler(int sig, siginfo_t *client, void *con)
 		mask = 0b10000000;
 		c = 0;
 	}
-	kill(client->si_pid, SIGUSR1);
+	if (sig == SIGUSR1 || sig == SIGUSR2)
+		kill(client->si_pid, SIGUSR1);
 }
 
 int	main(void)
